[Refactor] Hide MetalStructCompiler to the cpp file
