+++ taichi/codegen/codegen_metal.cpp
@@ -1,7 +1,8 @@
-#include "codegen_metal.h"
+#include "taichi/codegen/codegen_metal.h"
 
 #include <string>
-#include <taichi/ir/ir.h>
+
+#include "taichi/ir/ir.h"
 
 TLANG_NAMESPACE_BEGIN
 namespace metal {
